+++ ThousandSchnapsen.Simulator/Dto/InfoDto.cs
@@ -2,6 +2,6 @@ namespace ThousandSchnapsen.Simulator.Dto
 {
     public class InfoDto
     {
-        public int[] AvailableCardsIds { get; set; }
+        public byte[] AvailableCardsIds { get; set; }
     }
 }
